updated model

+++ lib/models/layers/horizontal_band.dart
@@ -26,7 +26,7 @@ class HorizontalBand extends Layer {
         type:
             (json['type'] as String).toLayerType() ?? LayerType.horizontalBand,
         value: json['value'] ?? 0.0,
-        allowedError: json['allowedError'] ?? 40.0,
+        allowedError: json['allowedError'].toDouble() ?? 40.0,
         color: colorFromJson(json['color']),
         isLocked: json['isLocked'] ?? false);
   }
