+++ lib/models/settings/axis_settings.dart
@@ -27,7 +27,7 @@ abstract class AxisSettings {
   static TextStyle textStyleFromJson(Map<String, dynamic> json) {
     return TextStyle(
       color: colorFromJson(json['color']),
-      fontSize: json['fontSize'],
+      fontSize: json['fontSize'].toDouble(),
       fontWeight: fontWeightFromJson(json['fontWeight']),
     );
   }
